enable depth testing

+++ source/mesh.cpp
@@ -115,7 +115,7 @@ compile_shader(std::filesystem::path& path, gl::GLenum shader_type) -> gl::GLuin
 {
     auto content                         = read_file_binary(path);
     std::array<char*, 1> shader_contents = { content.data() };
-    std::array<int, 1> shader_size       = { content.size() };
+    std::array<int, 1> shader_size       = { static_cast<int>(content.size()) };
 
     auto shader = gl::glCreateShader(shader_type);
